+++ src/line__parse/index.js
@@ -1,17 +1,18 @@
 /** @typedef {import('../_types').readable_stream_or_reader_T}readable_stream_or_reader_T */
 /** @typedef {import('./index.d.ts').line_iterator__on_line_T}line_iterator__on_line */
+/** @typedef {import('./index.d.ts').line__parse__params_T}line__parse__params_T */
 /**
  * @param {line_iterator__on_line|readable_stream_or_reader_T}on_line_or_readable_stream_or_reader
- * @param {readable_stream_or_reader_T|TextDecoder}[readable_stream_or_reader_or_text_decoder]
- * @param {TextDecoder}[text_decoder]
+ * @param {readable_stream_or_reader_T|line__parse__params_T}[readable_stream_or_reader_or_params]
+ * @param {line__parse__params_T}[params]
  * @returns {void|AsyncIterable<string>}
  * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/ReadableStreamDefaultReader/read#example_2_-_handling_text_line_by_line}
  * @private
  */
 export function line__parse(
 	on_line_or_readable_stream_or_reader,
-	readable_stream_or_reader_or_text_decoder,
-	text_decoder
+	readable_stream_or_reader_or_params,
+	params
 ) {
 	const on_line =
 		typeof on_line_or_readable_stream_or_reader === 'function'
@@ -20,10 +21,15 @@ export function line__parse(
 	/** @type {readable_stream_or_reader_T} */
 	const readable_stream_or_reader =
 		on_line
-		? readable_stream_or_reader_or_text_decoder
+		? readable_stream_or_reader_or_params
 		: on_line_or_readable_stream_or_reader
-	if (!on_line) text_decoder = readable_stream_or_reader_or_text_decoder
-	if (!text_decoder) text_decoder = new TextDecoder('utf-8')
+	if (!on_line) params = readable_stream_or_reader_or_params
+	const text_decoder =
+		params?.text_decoder
+		?? new TextDecoder('utf-8')
+	const include_line_separator =
+		params?.include_line_separator
+		?? false
 	return (
 		on_line
 		? line__parse__iterator_().next().then($=>$.value)
@@ -54,7 +60,9 @@ export function line__parse(
 				startIndex = re.lastIndex = 0
 				continue
 			}
-			const line = chunk.substring(startIndex, result.index)
+			const line = chunk.substring(
+				startIndex,
+				result.index + (include_line_separator ? 1 : 0))
 			if (on_line) {
 				on_line(line)
 			} else {